chore(webpack-cli): use new packages for better usability
