--- packages/webpack-cli/__tests__/command/external.test.js
@@ -1,9 +1,0 @@
-const ExternalCommand = require('../../lib/commands/external');
-
-describe('external command', () => {
-    it('should check existence of package', () => {
-        const exists = ExternalCommand.checkIfPackageExists('info');
-
-        expect(exists).toBeTruthy();
-    });
-});
