+++ packages/webpack-cli/bin/cli.js
@@ -4,7 +4,6 @@
 require('v8-compile-cache');
 const execa = require('execa');
 const importLocal = require('import-local');
-const logger = require('../lib/utils/logger');
 const parseArgs = require('../lib/utils/parse-args');
 
 // Prefer the local installation of webpack-cli
