+++ packages/webpack-cli/lib/utils/process-log.js
@@ -6,11 +6,13 @@ function logErrorAndExit(error) {
 }
 
 process.on('uncaughtException', error => {
+    console.log(logger);
     logger.error(`Uncaught exception: ${error}`);
     logErrorAndExit(error);
 });
 
 process.on('unhandledRejection', error => {
+    console.log(logger);
     logger.error(`Promise rejection: ${error}`);
     logErrorAndExit(error);
 });
