+++ packages/webpack-cli/package.json
@@ -38,7 +38,10 @@
     "interpret": "^2.0.0",
     "v8-compile-cache": "^2.1.0",
     "webpack-log": "^3.0.1",
-    "webpack-merge": "^4.2.2"
+    "webpack-merge": "^4.2.2",
+    "@webpack-cli/package-utils": "1.0.0",
+    "@webpack-cli/logger": "1.0.0",
+    "rechoir": "^0.7.0"
   },
   "peerDependencies": {
     "webpack": "^5.0.0-beta.12"
